+++ packages/app-starknet/src/index.ts
@@ -1,0 +1,8 @@
+export * from './app';
+export * from './types';
+export {
+  updateLogger,
+  setStarknetApiJs,
+  getAddressFromPublicKey,
+} from './utils';
+export type { starknetApiJsLibType } from './utils';