+++ apps/node/CHANGELOG.md
@@ -1,5 +1,16 @@
 # nodejs-test
 
+## 0.1.0
+
+### Minor Changes
+
+- [#137](https://github.com/Cypherock/sdk/pull/137) [`a680bb7`](https://github.com/Cypherock/sdk/commit/a680bb73a340e7c1ab60a88ba1555d0b8c4ad36f) Thanks [@vrockz747](https://github.com/vrockz747)! - Add starknet support
+
+### Patch Changes
+
+- Updated dependencies [[`a680bb7`](https://github.com/Cypherock/sdk/commit/a680bb73a340e7c1ab60a88ba1555d0b8c4ad36f)]:
+  - @cypherock/sdk-app-starknet@0.1.0
+
 ## 0.0.51
 
 ### Patch Changes